logic segregation
