+++ extern-api-java-sdk-integrational-tests/src/test/java/ru/kontur/extern_api/sdk/RelatedDocumentsIT.java
@@ -27,20 +27,13 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static ru.kontur.extern_api.sdk.utils.TestUtils.fromWin1251Bytes;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.List;
-import java.util.Objects;
-import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.apache.commons.io.IOUtils;
 import org.jetbrains.annotations.Contract;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
@@ -54,21 +47,17 @@ import ru.kontur.extern_api.sdk.model.Docflow;
 import ru.kontur.extern_api.sdk.model.DocflowPage;
 import ru.kontur.extern_api.sdk.model.DocflowType;
 import ru.kontur.extern_api.sdk.model.Document;
-import ru.kontur.extern_api.sdk.model.DocumentContents;
-import ru.kontur.extern_api.sdk.model.DocumentDescription;
 import ru.kontur.extern_api.sdk.model.DocumentType;
 import ru.kontur.extern_api.sdk.model.Draft;
-import ru.kontur.extern_api.sdk.model.ExtendedDraftMetaRequest;
 import ru.kontur.extern_api.sdk.model.InventoriesPage;
 import ru.kontur.extern_api.sdk.model.Inventory;
 import ru.kontur.extern_api.sdk.model.Signature;
 import ru.kontur.extern_api.sdk.model.TestData;
 import ru.kontur.extern_api.sdk.provider.crypt.mscapi.CryptoProviderMSCapi;
-import ru.kontur.extern_api.sdk.utils.Awaiter;
 import ru.kontur.extern_api.sdk.utils.CryptoUtils;
 import ru.kontur.extern_api.sdk.utils.DemandTestData;
 import ru.kontur.extern_api.sdk.utils.DemandTestDataProvider;
-import ru.kontur.extern_api.sdk.utils.Resources;
+import ru.kontur.extern_api.sdk.utils.RelatedDocflowProvider;
 import ru.kontur.extern_api.sdk.utils.SystemProperty;
 import ru.kontur.extern_api.sdk.utils.TestSuite;
 import ru.kontur.extern_api.sdk.utils.TestUtils;
@@ -76,21 +65,22 @@ import ru.kontur.extern_api.sdk.utils.TestUtils;
 @DisplayName("RelatedDocuments service should be able to")
 class RelatedDocumentsIT {
 
-    private static final String WINDOWS_1251 = "windows-1251";
-    private static final String VALID_INVENTORY_JPG_DOCUMENT = "/inventories/zapiska_valid.jpg";
+
     private static final String IFNS_CODE = "0087";
     private static ExternEngine engine;
     private static List<DemandTestData> tests;
     private static CryptoUtils cryptoUtils;
     private static int sentRelatedInventories;
     private static int sentRelatedLetters;
+    private static RelatedDocflowProvider relatedDocflowProvider;
 
     @BeforeAll
     static void setUpClass() {
         TestSuite testSuite = TestSuite.Load();
         engine = testSuite.engine;
         engine.setCryptoProvider(new CryptoProviderMSCapi());
-        cryptoUtils = CryptoUtils.with(engine.getCryptoProvider());
+        relatedDocflowProvider = new RelatedDocflowProvider(engine, IFNS_CODE);
+
         String certificate = cryptoUtils.loadX509(engine.getConfiguration().getThumbprint());
         TestData[] testData = TestUtils.getTestData(certificate);
 
@@ -333,225 +323,19 @@ class RelatedDocumentsIT {
                 demandAttachment.getDescription().getRelatedDocflowsCount());
     }
 
-
     private CompletableFuture<Docflow> sendRelatedLetter(DemandTestData testData) {
-        sentRelatedLetters++;
-        return engine.getRelatedDocumentsService(testData.getDemandId(), testData.getDemandAttachmentId())
-                .createRelatedDraft(new ExtendedDraftMetaRequest(TestUtils.toDraftMetaRequest(testData), "Письмо"))
-                .thenCompose(draft -> addLetterDocument(draft, testData))
-                .thenCompose(this::sendDraft)
-                .thenCompose(docflow -> waitForDocflow(docflow.getId()));
+        return relatedDocflowProvider.sendRelatedLetter(testData)
+                .thenApply(docflow -> {
+                    sentRelatedLetters++;
+                    return docflow;
+                });
     }
 
-    CompletionStage<Draft> addLetterDocument(Draft draft, DemandTestData testData) {
-        byte[] docContent = getContentBytes(getLetterData(testData));
-        return addAnyDocument(docContent, new DocumentDescription(), draft);
-    }
-
-    CompletableFuture<Inventory> sendRelatedInventory(DemandTestData testData) {
-        UUID childFileId = UUID.randomUUID();
-        UUID baseFileId = UUID.randomUUID();
-        sentRelatedInventories++;
-        return engine.getRelatedDocumentsService(testData.getDemandId(), testData.getDemandAttachmentId())
-                .createRelatedDraft(TestUtils.toDraftMetaRequest(testData))
-                .thenCompose(draft -> addInventoriesDocument(draft, testData, baseFileId, childFileId))
-                .thenCompose(draft -> addInventoriesSubmission(draft, testData, baseFileId, childFileId))
-                .thenCompose(this::sendDraft)
-                .thenCompose(docflow -> waitForInventory(docflow.getId(), testData));
-    }
-
-    private static CompletableFuture<Docflow> waitForDocflow(UUID id) {
-        return Awaiter.waitForCondition(() -> engine.getDocflowService()
-                        .lookupDocflowAsync(id)
-                        .thenApply(QueryContext::getDocflow),
-                Objects::nonNull, 2000);
-    }
-
-    private static CompletableFuture<Inventory> waitForInventory(UUID id, DemandTestData testData) {
-        return Awaiter.waitForCondition(
-                () -> engine.getRelatedDocumentsService(testData.getDemandId(), testData.getDemandAttachmentId())
-                        .getInventory(id),
-                Objects::nonNull, 2000);
-    }
-
-    String getAttachmentFileName(DemandTestData testData, UUID baseFileId, UUID childFileId) {
-        return "0510041_"
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "_"
-                + IFNS_CODE
-                + "_"
-                + baseFileId
-                + "_"
-                + childFileId
-                + ".jpg";
-    }
-
-    CompletableFuture<Draft> addInventoriesDocument(Draft draft, DemandTestData testData, UUID baseFileId,
-            UUID childFileId) {
-        byte[] docContent = getFromResource(VALID_INVENTORY_JPG_DOCUMENT);
-        DocumentDescription description = new DocumentDescription();
-        description.setFilename(getAttachmentFileName(testData, baseFileId, childFileId));
-
-        return addAnyDocument(docContent, description, draft);
-    }
-
-    CompletableFuture<Draft> addInventoriesSubmission(Draft draft, DemandTestData testData, UUID baseFileId,
-            UUID childFileId) {
-        byte[] docContent = getContentBytes(getSubmissionData(testData, baseFileId, childFileId));
-        DocumentDescription description = new DocumentDescription();
-        return addAnyDocument(docContent, description, draft);
-    }
-
-    private byte[] getContentBytes(String content) {
-        try {
-            return content.getBytes(WINDOWS_1251);
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    CompletableFuture<Draft> addAnyDocument(byte[] docContent, DocumentDescription description, Draft draft) {
-        DocumentContents contents = new DocumentContents();
-        contents.setBase64Content(toBase64(docContent));
-        contents.setSignature(getContentSignature(docContent));
-        contents.setDescription(description);
-        return engine.getDraftService().addDecryptedDocumentAsync(draft.getId(), contents)
-                .thenApply(result -> draft);
-    }
-
-    byte[] getFromResource(String path) {
-        try (InputStream is = Resources.class.getResourceAsStream(path)) {
-            Objects.requireNonNull(is, "Resource not found: " + path);
-            return IOUtils.toByteArray(is);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    String getContentSignature(byte[] docContent) {
-        return toBase64(cryptoUtils
-                .sign(engine.getConfiguration().getThumbprint(), docContent));
-    }
-
-    String toBase64(byte[] docContent) {
-        return new String(Base64.getEncoder().encode(docContent));
-    }
-
-    CompletableFuture<Docflow> sendDraft(Draft draft) {
-        return engine.getDraftService().sendAsync(draft.getId()).thenApply(QueryContext::get);
-    }
-
-
-    String getSubmissionData(DemandTestData testData, UUID baseFileId, UUID childFileId) {
-        return "<?xml version=\"1.0\" encoding=\"windows-1251\"?>"
-                + "<Файл ИдФайл=\"ON_DOCNPNO_" + IFNS_CODE + "_" + IFNS_CODE + "_"
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "_20180226_"
-                + UUID.randomUUID().toString()
-                + "\" ВерсПрог =\"КОНТУР-ЭКСТЕРН, ВЕРСИЯ 13.0\" ВерсФорм=\"5.01\">"
-                + "<Документ КНД=\"1184002\" ДатаДок=\"17.05.2018\">"
-                + "<СвОтпрДок>"
-                + "<ОтпрЮЛ НаимОрг=\"ООО «Первая Ижевская тестовая»\" ИННЮЛ=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + "\" КПП=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "\" />"
-                + "</СвОтпрДок>"
-                + "<СвПолДок>"
-                + "<ОтпрНО КодНО=\"" + IFNS_CODE + "\" НаимНО=\"Тестовый КО (положительный ответ)\" />"
-                + "</СвПолДок>"
-                + "<СвНП>"
-                + "<НПЮЛ НаимОрг=\"ООО «Первая Ижевская тестовая»\" ИННЮЛ=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + "\" КПП=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "\" />"
-                + "</СвНП>"
-                + "<Подписант ПрПодп=\"3\" Должн =\"Руководитель\" Тлф=\"8800\" E-mail=\"ruk@skbkontur.ru\" ИННФЛ=\"227632864503\">"
-                + "<ФИО Фамилия=\""
-                + testData.getClientInfo().getSender().getFio().getSurname()
-                + "\" Имя=\""
-                + testData.getClientInfo().getSender().getFio().getName()
-                + "\" Отчество=\""
-                + testData.getClientInfo().getSender().getFio().getPatronymic()
-                + "\" />"
-                + "</Подписант>"
-                + "<ДокПредстНО КолФайл=\"1\">"
-                + "<ИдФайлОсн>1165013_" + IFNS_CODE + "_"
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "_"
-                + UUID.randomUUID().toString()
-                + "_20170807_"
-                + UUID.randomUUID().toString()
-                + "</ИдФайлОсн>"
-                + "<ДокСкан ПорНомДок=\"1.01\" КодДок=\"0510041\" НаимДок=\"Фото_1\" >"
-                + "<ИмяФайл>"
-                + getAttachmentFileName(testData, baseFileId, childFileId)
-                + "</ИмяФайл>"
-                + "</ДокСкан>"
-                + "</ДокПредстНО>"
-                + "</Документ>"
-                + "</Файл>";
-    }
-
-
-    String getLetterData(DemandTestData testData) {
-        return "<?xml version=\"1.0\" encoding=\"windows-1251\"?>"
-                + "<Файл ИдФайл=\"IU_OBRNP_"
-                + IFNS_CODE
-                + "_"
-                + IFNS_CODE
-                + "_"
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "_20180301_"
-                + UUID.randomUUID().toString()
-                + "\" ВерсПрог=\"КОНТУР-ЭКСТЕРН, ВЕРСИЯ 13.0\" ВерсФорм=\"5.03\">"
-                + "<Документ КНД=\"1166102\" ДатаДок=\"01.03.2018\">"
-                + "<СвНП>"
-                + "<НПЮЛ НаимОрг=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getOrgName()
-                + "\" ИННЮЛ=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + "\" КПП=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "\"/>"
-                + "<АдрРФ КодРегион=\"66\"/>"
-                + "</СвНП>"
-                + "<СвОтпр>"
-                + "<СвОтпрЮЛ НаимОрг=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getOrgName()
-                + "\" ИННЮЛ=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getInn()
-                + "\" КПП=\""
-                + TestUtils.toDraftMetaRequest(testData).getPayer().getKpp()
-                + "\"/>"
-                + "</СвОтпр>"
-                + "<Подписант ПрПодп=\"1\">"
-                + "<ФИО Фамилия=\""
-                + testData.getClientInfo().getSender().getFio().getSurname()
-                + "\" Имя=\""
-                + testData.getClientInfo().getSender().getFio().getName()
-                + "\" Отчество=\""
-                + testData.getClientInfo().getSender().getFio().getPatronymic()
-                + "\" />"
-                + "</Подписант>"
-                + "<ОбращИнф ИФНС=\"" + IFNS_CODE + "\">"
-                + "<ОбращТекст><![CDATA[с]]></ОбращТекст>"
-                + "<Прил КолФайл=\"0\"/>"
-                + "</ОбращИнф>"
-                + "</Документ>"
-                + "</Файл>";
-    }
-
-    String fromWin1251Bytes(byte[] content) {
-        try {
-            return new String(content, WINDOWS_1251);
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("Win-1251 no supported");
-        }
+    private CompletableFuture<Inventory> sendRelatedInventory(DemandTestData testData) {
+        return relatedDocflowProvider.sendRelatedInventory(testData)
+                .thenApply(inventory -> {
+                    sentRelatedInventories++;
+                    return inventory;
+                });
     }
 }
