+++ extern-api-java-sdk-integrational-tests/src/test/java/ru/kontur/extern_api/sdk/utils/DemandTestData.java
@@ -5,6 +5,9 @@ import ru.kontur.extern_api.sdk.model.TestData;
 
 public class DemandTestData extends TestData {
 
+    private UUID demandId;
+    private UUID demandAttachmentId;
+
     public UUID getDemandId() {
         return demandId;
     }
@@ -13,18 +16,15 @@ public class DemandTestData extends TestData {
         this.demandId = demandId;
     }
 
-    private UUID demandId;
-    private UUID demandAttachmentId;
-
     public UUID getDemandAttachmentId() {
         return demandAttachmentId;
     }
 
-    public void setDemandAttachmentId(UUID demandAttachmentId) {
+    void setDemandAttachmentId(UUID demandAttachmentId) {
         this.demandAttachmentId = demandAttachmentId;
     }
 
-    public DemandTestData(TestData testData, UUID demandId) {
+    DemandTestData(TestData testData, UUID demandId) {
         this.demandId = demandId;
         this.setClientInfo(testData.getClientInfo());
         this.setDocs(testData.getDocs());
